adjust metrics plot, use undersampling for imbalance
