+++ workflow/scripts/metrics.py
@@ -16,29 +16,29 @@ with open(snakemake.input.label_encoder, "rb") as f:
     le = pickle.load(f)
 
 # get model results in loop
-df = pd.DataFrame()
+df_list = []
 for fold in range(0, snakemake.params.num_folds):
     for stage in ["train", "test"]:
-        file = snakemake.input[f"{stage}_labels"][fold]
-        with open(file, "rb") as f:
+        with open(snakemake.input[f"{stage}_labels"][fold], "rb") as f:
             y = pickle.load(f)
 
-        for model in snakemake.params.models:
+        file = [f for f in snakemake.input[f"{stage}_proba"]][fold]
+        with open(file, "rb") as f:
+            y_proba = pickle.load(f)
+
+        for label in ["KNRGL","RL1","OTHER"]:
             _df = pd.DataFrame()
-
-            file = [f for f in snakemake.input[f"{stage}_proba"] if model in f][fold]
-
-            with open(file, "rb") as f:
-                y_proba = pickle.load(f)[:, le.transform(["KNRGL"])[0]]
-
+            
             _df["precision"], _df["recall"], _ = precision_recall_curve(
-                y, y_proba, pos_label=le.transform(["KNRGL"])[0]
+                y, y_proba[:, le.transform([label])[0]], pos_label=le.transform([label])[0]
             )
-
+            
             _df["stage"] = stage
-            _df["model"] = model
+            _df["label"] = label
             _df["fold"] = fold + 1
-            df = pd.concat([df, _df]).reset_index(drop=True)
+            df_list.append(_df)
+        
+df = pd.concat(df_list).reset_index(drop=True)
 
 # Plot PR curves
 sns.set_style("ticks")
@@ -46,17 +46,17 @@ fig = sns.relplot(
     data=df,
     x="recall",
     y="precision",
-    hue="model",
+    hue="label",
     col="stage",
+    row="fold",
     kind="line",
     markers=True,
 )
 fig.set(
     ylim=[0, 1],
-    xlim=[0, 1],
-    title="Precision-Recall Curve: KNRGL vs. RL1, OTHER",
+    xlim=[0, 1]
 )
-plt.savefig(snakemake.output.prcurve, format="svg", dpi=300)
+plt.savefig(snakemake.output.prcurve, format="png", dpi=300)
 
 # TODO: compute confusion matrix
 # TODO: report feature importance
